requirements.txt
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.19.0
flask>=2.0.0
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.0.0 